+++ test/nebulex/adapters/multilevel_exclusive_test.exs
@@ -16,7 +16,6 @@ defmodule Nebulex.Adapters.MultilevelExclusiveTest do
 
   import Nebulex.CacheCase
 
-  alias Nebulex.Distributed.Cluster
   alias Nebulex.Distributed.TestCache.{Multilevel, MultilevelNil}
   alias Nebulex.Distributed.TestCache.Multilevel.{L1, L2, L3}
 
@@ -91,7 +90,7 @@ defmodule Nebulex.Adapters.MultilevelExclusiveTest do
 
   describe "distributed levels" do
     test "return cluster nodes" do
-      assert Cluster.pg_nodes(:multilevel_exclusive_l3) == [node()]
+      assert L3.nodes(:multilevel_exclusive_l3) == [node()]
     end
 
     test "joining new node" do
@@ -105,7 +104,7 @@ defmodule Nebulex.Adapters.MultilevelExclusiveTest do
         )
 
       # check cluster nodes
-      assert Cluster.pg_nodes(:multilevel_exclusive_l3) == [node, node()]
+      assert L3.nodes(:multilevel_exclusive_l3) == [node, node()]
 
       kv_pairs = for k <- 1..100, do: {k, k}